requirements.txt
python-telegram-bot==20.7
paramiko==3.4.0